+++ src/ts/credits.ts
@@ -1,7 +1,7 @@
-import { Credits } from '../types/Credits';
+import { Credit } from '../types/Credit';
 import { credits } from '../data/credits';
 
-const getCredits = async function (): Promise<Credits[]> {
+const getCredits = async function (): Promise<Credit[]> {
   return credits;
 };
