+++ src/types/Credit.ts
@@ -1,8 +1,8 @@
-interface Credits {
+interface Credit {
   name: string;
   icon: string;
   licence: string;
   text: string | null;
 }
 
-export type { Credits };
+export type { Credit };
