+++ src/types/Script.ts
